+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # SpiderKeeper Changelog
-##v0.2.0
+## v2.0.0
+- refactor
+- support py3
+- optimized api
+- optimized scheduler
+- more scalable (can support multiply spider service)
+## v0.2.0
 - support view job of multi daemons.
 - support run on multi daemons.
 - support choice running daemon automaticaly.
 
-##v0.1.1
+## v0.1.1
 - add status monitor(https://github.com/afaqurk/linux-dash)
 
-##v0.1.0
+## v0.1.0
 - initial.